nlp-service/requirements.txt
fastapi
uvicorn
pydantic
pillow
google-generativeai
python-dotenv